requirements.txt
# Core
numpy>=1.24.0,<3.0
torch>=2.0.0
pyyaml>=6.0
jinja2>=3.1.0
scikit-learn>=1.3.0
pandas>=2.0.0

# LLM clients (optional; install if using API backends)
requests>=2.28.0
httpx>=0.24.0

# Logging & experiment tracking
tensorboard>=2.14.0

# Evaluation & plotting
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0

# Dev / lint (optional for local dev)
ruff>=0.1.0
black>=23.0.0
mypy>=1.5.0
pytest>=7.4.0
